Refactor Volume Auto Grow to support additional volume types
This commit refactors the existing pgData volume auto grow code
to better support upcoming feature enhancements relating to auto
grow capability for the pgBackRest repository volume and pg_wal
volume.

Issue: PGO-2606
