+++ internal/controller/postgrescluster/postgres_test.go
@@ -13,21 +13,18 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/go-logr/logr/funcr"
 	"github.com/google/go-cmp/cmp/cmpopts"
 	volumesnapshotv1 "github.com/kubernetes-csi/external-snapshotter/client/v8/apis/volumesnapshot/v1"
 	"gotest.tools/v3/assert"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/crunchydata/postgres-operator/internal/controller/runtime"
 	"github.com/crunchydata/postgres-operator/internal/feature"
 	"github.com/crunchydata/postgres-operator/internal/initialize"
-	"github.com/crunchydata/postgres-operator/internal/logging"
 	"github.com/crunchydata/postgres-operator/internal/naming"
 	"github.com/crunchydata/postgres-operator/internal/postgres"
 	"github.com/crunchydata/postgres-operator/internal/testing/cmp"
@@ -848,318 +845,6 @@ volumeMode: Filesystem
 				})
 			})
 		})
-	})
-}
-
-func TestSetVolumeSize(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-	cluster := v1beta1.PostgresCluster{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "elephant",
-			Namespace: "test-namespace",
-		},
-		Spec: v1beta1.PostgresClusterSpec{
-			InstanceSets: []v1beta1.PostgresInstanceSetSpec{{
-				Name:     "some-instance",
-				Replicas: initialize.Int32(1),
-			}},
-		},
-	}
-
-	instance := &appsv1.StatefulSet{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "elephant-some-instance-wxyz-0",
-			Namespace: cluster.Namespace,
-		}}
-
-	setupLogCapture := func(ctx context.Context) (context.Context, *[]string) {
-		calls := []string{}
-		testlog := funcr.NewJSON(func(object string) {
-			calls = append(calls, object)
-		}, funcr.Options{
-			Verbosity: 1,
-		})
-		return logging.NewContext(ctx, testlog), &calls
-	}
-
-	// helper functions
-	instanceSetSpec := func(request, limit string) *v1beta1.PostgresInstanceSetSpec {
-		return &v1beta1.PostgresInstanceSetSpec{
-			Name: "some-instance",
-			DataVolumeClaimSpec: v1beta1.VolumeClaimSpec{
-				AccessModes: []corev1.PersistentVolumeAccessMode{corev1.ReadWriteOnce},
-				Resources: corev1.VolumeResourceRequirements{
-					Requests: map[corev1.ResourceName]resource.Quantity{
-						corev1.ResourceStorage: resource.MustParse(request),
-					},
-					Limits: map[corev1.ResourceName]resource.Quantity{
-						corev1.ResourceStorage: resource.MustParse(limit),
-					}}}}
-	}
-
-	desiredStatus := func(request string) v1beta1.PostgresClusterStatus {
-		desiredMap := make(map[string]string)
-		desiredMap["elephant-some-instance-wxyz-0"] = request
-		return v1beta1.PostgresClusterStatus{
-			InstanceSets: []v1beta1.PostgresInstanceSetStatus{{
-				Name:                "some-instance",
-				DesiredPGDataVolume: desiredMap,
-			}}}
-	}
-
-	t.Run("RequestAboveLimit", func(t *testing.T) {
-		recorder := events.NewRecorder(t, runtime.Scheme)
-		reconciler := &Reconciler{Recorder: recorder}
-		ctx, logs := setupLogCapture(ctx)
-
-		pvc := &corev1.PersistentVolumeClaim{ObjectMeta: naming.InstancePostgresDataVolume(instance)}
-		spec := instanceSetSpec("4Gi", "3Gi")
-		pvc.Spec = spec.DataVolumeClaimSpec.AsPersistentVolumeClaimSpec()
-
-		reconciler.setVolumeSize(ctx, &cluster, pvc, spec.Name)
-
-		assert.Assert(t, cmp.MarshalMatches(pvc.Spec, `
-accessModes:
-- ReadWriteOnce
-resources:
-  limits:
-    storage: 3Gi
-  requests:
-    storage: 3Gi
-`))
-		assert.Equal(t, len(*logs), 0)
-		assert.Equal(t, len(recorder.Events), 1)
-		assert.Equal(t, recorder.Events[0].Regarding.Name, cluster.Name)
-		assert.Equal(t, recorder.Events[0].Reason, "VolumeRequestOverLimit")
-		assert.Equal(t, recorder.Events[0].Note, "pgData volume request (4Gi) for elephant/some-instance is greater than set limit (3Gi). Limit value will be used.")
-	})
-
-	t.Run("NoFeatureGate", func(t *testing.T) {
-		recorder := events.NewRecorder(t, runtime.Scheme)
-		reconciler := &Reconciler{Recorder: recorder}
-		ctx, logs := setupLogCapture(ctx)
-
-		pvc := &corev1.PersistentVolumeClaim{ObjectMeta: naming.InstancePostgresDataVolume(instance)}
-		spec := instanceSetSpec("1Gi", "3Gi")
-
-		desiredMap := make(map[string]string)
-		desiredMap["elephant-some-instance-wxyz-0"] = "2Gi"
-		cluster.Status = v1beta1.PostgresClusterStatus{
-			InstanceSets: []v1beta1.PostgresInstanceSetStatus{{
-				Name:                "some-instance",
-				DesiredPGDataVolume: desiredMap,
-			}},
-		}
-
-		pvc.Spec = spec.DataVolumeClaimSpec.AsPersistentVolumeClaimSpec()
-
-		reconciler.setVolumeSize(ctx, &cluster, pvc, spec.Name)
-
-		assert.Assert(t, cmp.MarshalMatches(pvc.Spec, `
-accessModes:
-- ReadWriteOnce
-resources:
-  limits:
-    storage: 3Gi
-  requests:
-    storage: 1Gi
-	`))
-
-		assert.Equal(t, len(recorder.Events), 0)
-		assert.Equal(t, len(*logs), 0)
-
-		// clear status for other tests
-		cluster.Status = v1beta1.PostgresClusterStatus{}
-	})
-
-	t.Run("FeatureEnabled", func(t *testing.T) {
-		gate := feature.NewGate()
-		assert.NilError(t, gate.SetFromMap(map[string]bool{
-			feature.AutoGrowVolumes: true,
-		}))
-		ctx := feature.NewContext(ctx, gate)
-
-		t.Run("StatusNoLimit", func(t *testing.T) {
-			recorder := events.NewRecorder(t, runtime.Scheme)
-			reconciler := &Reconciler{Recorder: recorder}
-			ctx, logs := setupLogCapture(ctx)
-
-			pvc := &corev1.PersistentVolumeClaim{ObjectMeta: naming.InstancePostgresDataVolume(instance)}
-			spec := &v1beta1.PostgresInstanceSetSpec{
-				Name: "some-instance",
-				DataVolumeClaimSpec: v1beta1.VolumeClaimSpec{
-					AccessModes: []corev1.PersistentVolumeAccessMode{corev1.ReadWriteOnce},
-					Resources: corev1.VolumeResourceRequirements{
-						Requests: map[corev1.ResourceName]resource.Quantity{
-							corev1.ResourceStorage: resource.MustParse("1Gi"),
-						}}}}
-			cluster.Status = desiredStatus("2Gi")
-			pvc.Spec = spec.DataVolumeClaimSpec.AsPersistentVolumeClaimSpec()
-
-			reconciler.setVolumeSize(ctx, &cluster, pvc, spec.Name)
-
-			assert.Assert(t, cmp.MarshalMatches(pvc.Spec, `
-accessModes:
-- ReadWriteOnce
-resources:
-  requests:
-    storage: 1Gi
-`))
-			assert.Equal(t, len(recorder.Events), 0)
-			assert.Equal(t, len(*logs), 0)
-
-			// clear status for other tests
-			cluster.Status = v1beta1.PostgresClusterStatus{}
-		})
-
-		t.Run("LimitNoStatus", func(t *testing.T) {
-			recorder := events.NewRecorder(t, runtime.Scheme)
-			reconciler := &Reconciler{Recorder: recorder}
-			ctx, logs := setupLogCapture(ctx)
-
-			pvc := &corev1.PersistentVolumeClaim{ObjectMeta: naming.InstancePostgresDataVolume(instance)}
-			spec := instanceSetSpec("1Gi", "2Gi")
-			pvc.Spec = spec.DataVolumeClaimSpec.AsPersistentVolumeClaimSpec()
-
-			reconciler.setVolumeSize(ctx, &cluster, pvc, spec.Name)
-
-			assert.Assert(t, cmp.MarshalMatches(pvc.Spec, `
-accessModes:
-- ReadWriteOnce
-resources:
-  limits:
-    storage: 2Gi
-  requests:
-    storage: 1Gi
-`))
-			assert.Equal(t, len(recorder.Events), 0)
-			assert.Equal(t, len(*logs), 0)
-		})
-
-		t.Run("BadStatusWithLimit", func(t *testing.T) {
-			recorder := events.NewRecorder(t, runtime.Scheme)
-			reconciler := &Reconciler{Recorder: recorder}
-			ctx, logs := setupLogCapture(ctx)
-
-			pvc := &corev1.PersistentVolumeClaim{ObjectMeta: naming.InstancePostgresDataVolume(instance)}
-			spec := instanceSetSpec("1Gi", "3Gi")
-			cluster.Status = desiredStatus("NotAValidValue")
-			pvc.Spec = spec.DataVolumeClaimSpec.AsPersistentVolumeClaimSpec()
-
-			reconciler.setVolumeSize(ctx, &cluster, pvc, spec.Name)
-
-			assert.Assert(t, cmp.MarshalMatches(pvc.Spec, `
-accessModes:
-- ReadWriteOnce
-resources:
-  limits:
-    storage: 3Gi
-  requests:
-    storage: 1Gi
-`))
-
-			assert.Equal(t, len(recorder.Events), 0)
-			assert.Equal(t, len(*logs), 1)
-			assert.Assert(t, cmp.Contains((*logs)[0], "Unable to parse volume request: NotAValidValue"))
-		})
-
-		t.Run("StatusWithLimit", func(t *testing.T) {
-			recorder := events.NewRecorder(t, runtime.Scheme)
-			reconciler := &Reconciler{Recorder: recorder}
-			ctx, logs := setupLogCapture(ctx)
-
-			pvc := &corev1.PersistentVolumeClaim{ObjectMeta: naming.InstancePostgresDataVolume(instance)}
-			spec := instanceSetSpec("1Gi", "3Gi")
-			cluster.Status = desiredStatus("2Gi")
-			pvc.Spec = spec.DataVolumeClaimSpec.AsPersistentVolumeClaimSpec()
-
-			reconciler.setVolumeSize(ctx, &cluster, pvc, spec.Name)
-
-			assert.Assert(t, cmp.MarshalMatches(pvc.Spec, `
-accessModes:
-- ReadWriteOnce
-resources:
-  limits:
-    storage: 3Gi
-  requests:
-    storage: 2Gi
-`))
-			assert.Equal(t, len(recorder.Events), 0)
-			assert.Equal(t, len(*logs), 0)
-		})
-
-		t.Run("StatusWithLimitGrowToLimit", func(t *testing.T) {
-			recorder := events.NewRecorder(t, runtime.Scheme)
-			reconciler := &Reconciler{Recorder: recorder}
-			ctx, logs := setupLogCapture(ctx)
-
-			pvc := &corev1.PersistentVolumeClaim{ObjectMeta: naming.InstancePostgresDataVolume(instance)}
-			spec := instanceSetSpec("1Gi", "2Gi")
-			cluster.Status = desiredStatus("2Gi")
-			pvc.Spec = spec.DataVolumeClaimSpec.AsPersistentVolumeClaimSpec()
-
-			reconciler.setVolumeSize(ctx, &cluster, pvc, spec.Name)
-
-			assert.Assert(t, cmp.MarshalMatches(pvc.Spec, `
-accessModes:
-- ReadWriteOnce
-resources:
-  limits:
-    storage: 2Gi
-  requests:
-    storage: 2Gi
-`))
-
-			assert.Equal(t, len(*logs), 0)
-			assert.Equal(t, len(recorder.Events), 1)
-			assert.Equal(t, recorder.Events[0].Regarding.Name, cluster.Name)
-			assert.Equal(t, recorder.Events[0].Reason, "VolumeLimitReached")
-			assert.Equal(t, recorder.Events[0].Note, "pgData volume(s) for elephant/some-instance are at size limit (2Gi).")
-		})
-
-		t.Run("DesiredStatusOverLimit", func(t *testing.T) {
-			recorder := events.NewRecorder(t, runtime.Scheme)
-			reconciler := &Reconciler{Recorder: recorder}
-			ctx, logs := setupLogCapture(ctx)
-
-			pvc := &corev1.PersistentVolumeClaim{ObjectMeta: naming.InstancePostgresDataVolume(instance)}
-			spec := instanceSetSpec("4Gi", "5Gi")
-			cluster.Status = desiredStatus("10Gi")
-			pvc.Spec = spec.DataVolumeClaimSpec.AsPersistentVolumeClaimSpec()
-
-			reconciler.setVolumeSize(ctx, &cluster, pvc, spec.Name)
-
-			assert.Assert(t, cmp.MarshalMatches(pvc.Spec, `
-accessModes:
-- ReadWriteOnce
-resources:
-  limits:
-    storage: 5Gi
-  requests:
-    storage: 5Gi
-`))
-
-			assert.Equal(t, len(*logs), 0)
-			assert.Equal(t, len(recorder.Events), 2)
-			var found1, found2 bool
-			for _, event := range recorder.Events {
-				if event.Reason == "VolumeLimitReached" {
-					found1 = true
-					assert.Equal(t, event.Regarding.Name, cluster.Name)
-					assert.Equal(t, event.Note, "pgData volume(s) for elephant/some-instance are at size limit (5Gi).")
-				}
-				if event.Reason == "DesiredVolumeAboveLimit" {
-					found2 = true
-					assert.Equal(t, event.Regarding.Name, cluster.Name)
-					assert.Equal(t, event.Note,
-						"The desired size (10Gi) for the elephant/some-instance pgData volume(s) is greater than the size limit (5Gi).")
-				}
-			}
-			assert.Assert(t, found1 && found2)
-		})
-
 	})
 }
